refactor: rename component to ReceiptListRedirect
